[ALGORITHM] Knuth-Morris-Pratt algorithm
Added an implementation of the KMP algorithm. It solves the question
what is the longest proper prefix that is also a proper suffix of a given
string? that I had. It allows to find all ocurrences of a pattern in a text
in O(|text|) with a preparation step of O(|pattern|).